issue/12 Simplfy utilties.py if/elif/else and move test fixture to conftest.py
